+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Features
 
 ### Fixes
+- **ElementMetadata consolidation** Now `text_as_html` metadata is combined across all elements in CompositeElement when chunking HTML output
 
 ## 0.16.5